# tests/snapshots/openapi_v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MeResponse": {
        "additionalProperties": false,
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          },
          "last_login_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "status",
          "is_admin",
          "created_at"
        ],
        "title": "MeResponse",
        "type": "object"
      },
      "MeSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status",
          "updated_at"
        ],
        "title": "MeSessionResponse",
        "type": "object"
      },
      "MeSessionsResponse": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/MeSessionResponse"
            },
            "title": "Sessions",
            "type": "array"
          }
        },
        "required": [
          "page",
          "page_size",
          "has_more"
        ],
        "title": "MeSessionsResponse",
        "type": "object"
      },
      "MonitoringOverviewResponse": {
        "additionalProperties": false,
        "properties": {
          "average_recommendation_count": {
            "title": "Average Recommendation Count",
            "type": "number"
          },
          "completion_rate": {
            "title": "Completion Rate",
            "type": "number"
          },
          "duration_p50_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration P50 Ms"
          },
          "duration_p95_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration P95 Ms"
          },
          "failure_rate": {
            "title": "Failure Rate",
            "type": "number"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "implicit_usage_rate": {
            "title": "Implicit Usage Rate",
            "type": "number"
          },
          "jd_evidence_coverage_rate": {
            "title": "Jd Evidence Coverage Rate",
            "type": "number"
          },
          "reordered_run_count": {
            "title": "Reordered Run Count",
            "type": "integer"
          },
          "stage_latencies": {
            "items": {
              "$ref": "#/components/schemas/StageLatencyResponse"
            },
            "title": "Stage Latencies",
            "type": "array"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Status Counts",
            "type": "object"
          },
          "total_runs": {
            "title": "Total Runs",
            "type": "integer"
          },
          "warning_rate": {
            "title": "Warning Rate",
            "type": "number"
          },
          "window_hours": {
            "title": "Window Hours",
            "type": "integer"
          }
        },
        "required": [
          "window_hours",
          "generated_at",
          "total_runs",
          "completion_rate",
          "warning_rate",
          "failure_rate",
          "average_recommendation_count",
          "jd_evidence_coverage_rate",
          "implicit_usage_rate",
          "reordered_run_count"
        ],
        "title": "MonitoringOverviewResponse",
        "type": "object"
      },
      "OtpRequest": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "enum": [
              "email",
              "phone"
            ],
            "title": "Channel",
            "type": "string"
          },
          "target": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "target"
        ],
        "title": "OtpRequest",
        "type": "object"
      },
      "OtpRequestAccepted": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "const": "otp_accepted",
            "default": "otp_accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "OtpRequestAccepted",
        "type": "object"
      },
      "OtpVerifyRequest": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "enum": [
              "email",
              "phone"
            ],
            "title": "Channel",
            "type": "string"
          },
          "code": {
            "pattern": "^\\d{6}$",
            "title": "Code",
            "type": "string"
          },
          "target": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "target",
          "code"
        ],
        "title": "OtpVerifyRequest",
        "type": "object"
      },
...
      "ProfilePatchRequest": {
        "additionalProperties": false,
        "properties": {
          "profile": {
            "additionalProperties": true,
            "title": "Profile",
            "type": "object"
          }
        },
        "required": [
          "profile"
        ],
        "title": "ProfilePatchRequest",
        "type": "object"
      },
      "ProfileResponse": {
        "additionalProperties": false,
        "properties": {
          "profile": {
            "additionalProperties": true,
            "title": "Profile",
            "type": "object"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "updated_at"
        ],
        "title": "ProfileResponse",
        "type": "object"
      },
...
      "ReactionRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "job_id": {
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "outcome": {
            "minLength": 1,
            "title": "Outcome",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "user_rating": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Rating"
          }
        },
        "required": [
          "job_id",
          "outcome"
        ],
        "title": "ReactionRequest",
        "type": "object"
      },
      "ReactionResponse": {
        "additionalProperties": false,
        "properties": {
          "feedback_id": {
            "title": "Feedback Id",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "const": "reaction_recorded",
            "default": "reaction_recorded",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "feedback_id"
        ],
        "title": "ReactionResponse",
        "type": "object"
      },
      "RecentRunResponse": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "recommendation_count": {
            "default": 0,
            "title": "Recommendation Count",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warning_codes": {
            "items": {
              "type": "string"
            },
            "title": "Warning Codes",
            "type": "array"
          }
        },
        "required": [
          "run_id",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "RecentRunResponse",
        "type": "object"
      },
      "RecentRunsResponse": {
        "additionalProperties": false,
        "properties": {
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RecentRunResponse"
            },
            "title": "Runs",
            "type": "array"
          },
          "window_hours": {
            "title": "Window Hours",
            "type": "integer"
          }
        },
        "required": [
          "window_hours",
          "generated_at"
        ],
        "title": "RecentRunsResponse",
        "type": "object"
      },
...
      "StageLatencyResponse": {
        "additionalProperties": false,
        "properties": {
          "p50_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "P50 Ms"
          },
          "p95_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Ms"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage"
        ],
        "title": "StageLatencyResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Career-RAG",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/otp/request": {
      "post": {
        "operationId": "request_otp_api_v1_auth_otp_request_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpRequestAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Otp"
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "operationId": "verify_login_otp_api_v1_auth_otp_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Login Otp"
      }
    },
...
    "/api/v1/me/profile": {
      "get": {
        "operationId": "get_profile_api_v1_me_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Profile"
      },
      "patch": {
        "operationId": "patch_profile_api_v1_me_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Profile"
      }
    },
    "/api/v1/me/sessions": {
      "get": {
        "operationId": "get_my_sessions_api_v1_me_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeSessionsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get My Sessions"
      }
    },
    "/api/v1/monitoring/overview": {
      "get": {
        "operationId": "monitoring_overview_api_v1_monitoring_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_hours",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 720,
              "minimum": 1,
              "title": "Window Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoringOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Monitoring Overview"
      }
    },
    "/api/v1/monitoring/runs": {
      "get": {
        "operationId": "monitoring_runs_api_v1_monitoring_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "window_hours",
            "required": false,
            "schema": {
              "default": 24,
              "maximum": 720,
              "minimum": 1,
              "title": "Window Hours",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecentRunsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Monitoring Runs"
      }
    },
...
    "/api/v1/runs/{run_id}/reaction": {
      "post": {
        "operationId": "add_run_reaction_api_v1_runs__run_id__reaction_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Run Reaction"
      }
    },
...
  }
}